Land Recent QUIC changes.
Addressing comments in Jana's review of cr/52231261.

Merge internal change: 53582401

Add a temporary legacy quic constructor

Merge internal change: 53506960

QUIC: disable P-256 support on the server.

The P-256 key generation is done with OpenSSL, which doesn't use the
QuicRandom passed to DefaultConfig(). This is causing the generated
server configs to be non-deterministic and breaking 0-RTT handshakes.

Merge internal change: 53501783

Fix an LOG to use the correct condition in QuicReceivedPacketManager and
change it to a DFATAL so in the future tests will prevent re-occurrence.

Merge internal change: 53483753

Cleanup: Rename OnIncomingAck to OnPacketAcked, and remove unneeeded
argument from SentPacketManager::OnIncomingAck.

Merge internal change: 53483155

Fix a bug in QuicConnection/QuicConnectionHelper if the helper buffered
the write (as is the case in chrome).  In this case, the sent packet was
not accounted for properly.

Merge internal change: 53462749

Refactor to change WritePacket to return a WriteResult struct.

Merge internal change: 53382221

Fixing a bug where the version negotiation packet would get dropped on
the floor if the socket was write blocked at the time it was sent. The
packet is now queued.

Merge internal change: 53317846

Create a new QUIC_INVALID_CHANNEL_ID_SIGNATURE error to replace a usage
of QUIC_INTERNAL_ERROR.

Merge internal change: 53277933

Added a todo to merge internal CL 53267501 when chromium's version of
OpenSSL has latest AEAD code.

Didn't merge internal change: 53267501

R=rch@chromium.org

Review URL: https://codereview.chromium.org/26385004