Do not send mouse-move when target is destroyed due to synthesized mouse-exit in RootWindow
When a RootWindow is going to dispatch a mouse-move to a child, if the
target is different than the current mouse_moved_handler_, it sends a
synthesized mouse-exit to to mouse_moved_handler_. The target might be
destroyed because of this mouse-exit. If that's the case, the original
mouse-move should not be sent to the destroyed target.

BUG=304276

Review URL: https://codereview.chromium.org/26499003