Add chromium-committers appengine app.
This app hosts the lists of all Chromium committers and will become the
golden source of truth for that list. It receives updates from an internal
google service. It displays the list only to logged-in users whose login email
is found in the list itself, or other apps that successfully authenticate with
it via hmac (such as the commit queue).

Review URL: https://codereview.chromium.org/25515004