Check the extension for experimental API permission.
The current code that checks to see if the extension is enabled for
Experimental API calls only checks the command-line flags. It needs
to also check to see if the Extension has been granted Experimental
API permission from the webstore.

The old command line check is still needed to detect when the flag
is set directly in chrome://flags.

BUG=305342
R=scheib@chromium.org

Review URL: https://codereview.chromium.org/26756002