Improve perf of CalculateDrawPropertiesInternal when there are no scroll parents.
When we have scroll parents/children, we have to, potentially, reorder the children.
Previously, we had incurred a portion of this cost, even if we didn't have any scroll
parents/children in the tree. We would still populate a vector of pointers to the
children (representing the order that they would be visited in). Of course, this
order was the same as visiting layer->children() in order, so this work is wasted.
This patch stores a bool on a layer's draw properties indicating that it has a
child with a scroll parent. If this value is false, the reordering work may be
safely skipped.

BUG=305565

Review URL: https://codereview.chromium.org/26685002