Do not add extension items to panel's context menu if extension is not present.
When the panel is navigated to a url outside of the extension's package, the
RenderViewHost isn't associated with the extension, so it will return NULL when
calling GetExtensionForRenderViewHost. This will cause a crash when
right-clicking a panel or detached panel (with a url outside of the extension's
page).

BUG=168998
TEST=1. Please refer to the crbug.com/168998 and load the attached extension in
crbug.com/168998#c4 .

Review URL: https://codereview.chromium.org/25705002