Deal correctly with too-small MediaCodec buffers.
Prior to r221331, MediaCodecBridge::QueueInputBuffer() returned the number of
bytes queued in an attempt to deal with clients that tried to queue inputs
larger than MediaCodec's buffers could hold.  That wasn't completely worked
through, though, and doesn't seem to be workable given that MediaCodec's API
deals in "Access Units", not bytes, so it is not permitted to queue half of an
input in one buffer and half in the next.  r221331 (silently) replaced this
half-support with a requirement that inputs fit in buffers.  This CL turns this
DCHECK requirement into an error-triggering one, and fixes a callsite that r221331 missed
(AndroidVideoDecodeAccelerator), along with a bunch of plumbing that was only
needed while partial-buffer-queueing was possible.

BUG=305421

Review URL: https://codereview.chromium.org/26630002