Cast: Cleaning up framer code; resolving TODOs
// TODO(pwestin): we can deprecate the timeout argument.
The timeout argument passed to 
Framer::GetEncodedVideoFrame and Framer::GetEncodedAudioFrame is removed since is was not used.

// TODO(pwestin): only release if we have a complete frame
VideoReceiver::GetEncodedVideoFrame and AudioReceiver::PlayoutTimeout()
Update logic such that a frame is only checked if it can be released if the incoming packet make a frame complete.

Review URL: https://codereview.chromium.org/26686004