[sync] Significantly speed up password model association on mac
During password model association on mac, the helper function
internal_keychain_helpers::GetPasswordsForForms individually searches
through the OS keychain for passwords matching each web form for which a
password was saved. With tracing turned on, it turned out that these
O(N) keychain searches, one for each password stored in the keychain,
accounted for more than 80% of the password model association time.

In this patch, we make GetPasswordsForForms first pre-load the attributes
of all passwords stored in the keychain without loading the actual password
data (this doesn't require authorization and won't pop up a UI prompt), and
then individually load password data only for the keychain entries that match
the web forms we're interested in.

This is still considerably faster than doing O(N) keychain searches.
For a test account with 50 synced passwords, this change yielded almost
an 80% reduction in the time spent in GetPasswordsForForms().

BUG=263685
TEST=Enable tracing for password model association and measure its
performance with and without this change.

R=stuartmorgan@chromium.org

Review URL: https://codereview.chromium.org/23477015