Hide touch selection cursor handle when out of client view
Selection handles are made hidden when they are out of client view. This
was not done for cursor handle. Other than fixing this, a check is added
to move the handles only when they're going to be visible. Since hiding
a handle fades it out, it is better to have this fade out in the
previous handle location which is inside the boundaries of the client
view.

BUG=289014

Review URL: https://codereview.chromium.org/25609006