Pepper: Remove in-process specific FileIO stuff.
RequestOSFileHandle allows a FileHandle to be provided to an in-process plugin,
regardless of quota policy. I believe this exists to support FileDownloader in
the NaCl trusted plugin, which will only ever open a file with read permission.

It's more straightforward to not impose any quota restrictions if the file is
only opened for reading, and removes the in-process check.

BUG=246396
R=dmichael@chromium.org

Review URL: https://codereview.chromium.org/26896003