[NaCl SDK] nacl_io: fix handling of missing interfaces.
Previously, for example, TCPInterface() would return
a valid interface pointer, even when the TCP interface
was missing.

This meant that the checks we had in place for the
precense of a given inteface were not working, and
calls to methods on the interface would crash.

R=noelallen@chromium.org

Review URL: https://codereview.chromium.org/26888003