Don't kill the GPU process on Chrome OS if we're not on VT1
When Chrome OS is switched to VT2, the GPU process watchdog will
eventually kick in and kill the GPU process. This would be fine,
as this is only possible in dev mode, except for the fact that
some people run dev mode and still report their crashes. This
creates a lot of noise in the crash reports, which makes finding
the real issues difficult.

To fix this, we modify the GPU watchdog to check the active tty
before killing the GPU process, and if the tty isn't 1, then we
will leave it alone, assuming that we are simply on a dev-mode
machine.

BUG=chromium:222224
TEST=switch to VT2, wait 10 seconds, switch back: the GPU process didn't get killed
TEST=crash the GPU while on VT1, the watchdog still triggers

Review URL: https://codereview.chromium.org/25037011