Treat all coordinates to RWHVA::SetBounds() as screen coordinates.
This is a speculative re-upload, in the hopes that random crashes we are seeing on the ChromeOS waterfall will have been fixed by Issue 268648 (issue is not re-producible anywhere but the waterfall).  If CrOS waterfall begins experiencing memory corruption errors, this change should be reverted.

BUG=176221
TBR=sky

Review URL: https://codereview.chromium.org/24901003