cc: Reset resource update controller on closing of layer tree host
This fixes a potential use-after-free bug.

The ResourceUpdateController might have posted a task using a weak
pointer to itself. That task can result in a commit, which assumes
a number of resources are still available on the LayerTreeHostImpl.

However, ThreadProxy::LayerTreeHostClosedOnImplThread might have
freed those resources before the ThreadProxy itself is destroyed,
allowing the ResourceUpdateController's posted task to run without
the resources available.

This patch resets the ResourceUpdateController in
ThreadProxy::LayerTreeHostClosedOnImplThread now to avoid the race
between the posted task and the destruction of the ThreadProxy.

BUG=303828

Review URL: https://codereview.chromium.org/26522004