Add DOM distiller service
This adds the core DomDistillerService, a wrapped version of that
making it a BrowserContextKeyedService, and the corresponding
BrowserContextKeyedServiceFactory.

Currently, the service has an interface but doesn't actually implement
any of the methods.

This also adds some interfaces to the distiller that the service 
depends on. These also have no implementation.

BUG=288015

Review URL: https://codereview.chromium.org/26338002