webui: Allow to show empty menus and let them ignore key operations.
Previously we use to prevent menu buttons from showing an empty menu because
sending key events to an empty menu causes to freeze. But we use an custom empty
menu in the new tab page. This CL lets menu buttons allow to show an empty menu
and lets the menu ignore the key operations to avoid the freeze.

BUG=305585
TEST=manually

Review URL: https://codereview.chromium.org/26789002