base: Include intrin.h to get _ReturnAddress from clang
Clang provides an implementation of _ReturnAddress in intrin.h, and you
have to include the header to get it.  Without the include, we get link
errors because _ReturnAddress looks like a regular declaration.

According to the comments, this manual declaration was added to work
around a compiler problem in VS2005, but Chrome no longer supports that
toolchain.

R=cpu@chromium.org
BUG=82385

Review URL: https://codereview.chromium.org/26863006