C++ bindings
This CL includes code for a static library that supports C++ bindings to the Mojo message format. A sample interface with "generated" bindings is included in the sample/ folder along with sample_test.cc that shows how the generated code may be used.

This is a WIP but seems worth snapshotting as is. See the TODO file for a list of near term improvements, most notably testing.

Review URL: https://codereview.chromium.org/23913008