Implement first part of supporting global extension commands.
This first part simply reads the optional "global" flag from the manifest and registers the shortcut as global if it is set to true. The intent is to let developers register Ctrl+Shift+[0..9] (and nothing else) as global shortcuts but let users override that -- delete the shortcut, change it to whatever other combination (not restricted to Ctrl+Shift+[0..9] or even make it non-global. There's no UI at the moment to do so, but that is a separate CL that is almost ready.

This CL implements this functionality for Windows and stubs out the other platforms. The GTK stub has been fleshed out in parallel, but that's also another CL.

BUG=302437
R=erg@chromium.org, sky@chromium.org, thakis@chromium.org, yoz@chromium.org

Review URL: https://codereview.chromium.org/23812010