Add a flag for enabling/disabling the new accelerated scrolling path
We now have two accelerated scrolling paths. One accelerates scrolling whenever
it's safe to promote the overflow scrolling div to a stacking container /
containing block and not break stacking or clipping.

The new path, the 'universal' path, enables accelerated scrolling in those cases
where the old path would not opt-in.

Eventually, we'd like to eliminate the old path, but this will require layer
squashing (the new path can result in many composited layers that ought to be
combined -- i.e., a 'layer explosion'). Until these paths are mature, we need to
be able to enable and disable them both, so we have two separate sets of flags.

Once we've eliminated the old path, we can replace both with
--[enable/disable]-accelerated-overflow-scroll (and eventually get rid of both
sets of flags).

BUG=None

Review URL: https://codereview.chromium.org/26809004