Enable linear fade scrollbar animation on programmatic scroll
The result of ScrollbarAnimationControllerLinearFade::DidScrollUpdate was
reversed, depending on whether the update was programmatic or not. Return
true when the update was programmatic, forcing the animation. This fixes
an issue where the page sets the scroll offset initially resulting in no
fade animation.

NOTRY=true
BUG=306638

Review URL: https://codereview.chromium.org/26259008