Fix race conditions in window snapshot code.
Use LatencyInfo to track when a rendered frame reaches the screen.
Only tested so far on Linux with single-threaded compositor. Must be
tested with Telemetry's GpuTabTest.testScreenshot test on all of
Windows, Mac and Linux with the compositor disabled, enabled and
single-threaded, and enabled and multi-threaded.

BUG=256848

Review URL: https://codereview.chromium.org/23694031