Clean up the policy templates file.
The ChromeOS platform support versions were written both as 0.NN and NN;
changed to always use NN as for the other platforms.

The script that parses this file is currently ignoring the concrete version
number and only looks for a '-', and doesn't need to be updated.

BUG=None

Review URL: https://codereview.chromium.org/26333006