Implement linux part of global command, including linux_aura and gtk+.
Xlib API: XGrabKey/XUngrabKey are used in the stuff, but there is a subtle
difference between them. In gtk+, a event filter is added to the X root window
to intercept KeyPress event, but in linux_aura, a dispatcher is added to the
UI message loop.

Also note that because XGrabKey only works on the exact modifiers mask, we
should register our hot keys with modifiers that we want to ignore, including
Num lock, Caps lock, Scroll lock.

BUG=302437

Review URL: https://codereview.chromium.org/26538012