Enable chromium trace when atrace is started.
The X event change (https://codereview.chromium.org/23799005/) depends
on the TraceBuffer to handle X events, so we need to enable chromium
trace when using systrace on Android to trace WebView or Chrome.

Part of this change is separated from https://codereview.chromium.org/26675003/
which is not ready to submit before WebView and Android are ready for
combined trace.

BUG=trace-viewer:142

Review URL: https://codereview.chromium.org/26848006