[Android] Hide handles when the containing view moves
We temporarily hide handles when there is a scroll or zoom in the
content area. When the containing view is moved, we should also hide
the handles. This is particularly important for webview where scrolling
is often handled outside of the webview itself (and is sometimes mixed
between the two).

BUG=b/10892868

Review URL: https://codereview.chromium.org/26575006