Move renderer plugin code into a new component (re-land)
Android Webview needs to use some of the renderer plugin code, however
this code is currently in chrome. This commit moves the parts of this
code that don't depend on the rest of chrome into a new component.

This is a new commit of https://codereview.chromium.org/23606022
which had to be reverted due to a use after free problem, now
fixed.

The only change from the previous CL is to
chrome_plugin_placeholder.cc so TBR'ing owners of other files.

TBR=darin@chromium.org, joi@chromium.org
BUG=283713, 306815, 306803

Review URL: https://codereview.chromium.org/27197004