[MediaGalleries] Fix device id passing from iapps finder.
Note that in non-debug builds, there may be profile corruption with an invalid device ID stored in the "media_galleries" section of preferences. This will cause crashes or broken behavior elsewhere in media galleries.

The fix is to either wipe the profile or edit that line out of the file.

R=vandebo@chromium.org
BUG=307825

Review URL: https://codereview.chromium.org/27472005