Make sure the scheduler is READY_TO_COMMIT
Adds a DCHECK() to make sure that commit_state_ is what we expect
in the scheduler state machine.

Ran compositor performance and unit tests in debug on Android, visited
some stress-test sites in debug on Linux to make sure it's not getting
commonly hit.

BUG=308006
R=brianderson@chromium.org

Review URL: https://codereview.chromium.org/27526002