Don't clone WebAnimation in WebLayerImpl::addAnimation().
Blink does not need a WebAnimation object after passing it to Chromium, so we
can pass the ownership of cc:Animation in WebAnimation to Chromium compositor
instead of cloning.

ANIMATION_OWNERSHIP_TRANSFER is a temporary macro to change addAnimation() in
both Blink and Chromium. If ANIMATION_OWNERSHIP_TRANSFER is defined,
Blink passes ownership of WebAnimation to Chromium, so Chromium must delete the
given pointer.

BUG=290217

Review URL: https://codereview.chromium.org/24165003