Linux GYP generator for GN.
The Linux generator needs to be different than the Windows one since it supports cross-compiles. However, Windows doesn't support the conditions based on the toolset.

This doesn't actually hook up the cross-compile yet. Currently it passes null for the host targets. I'll do that in a future pass.

To make this formatted correctly, I redid how the indenting worked in the GYP output code.

R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/27563003