Revert "Revert "Allow clipping by scroll parents." and "Improve perf of CalculateDrawPropertiesInternal when there are no scroll parents.""
This reverts commit 97ada322de9f492e3f0d1741ac117fc42599451b.

This was a speculative perf revert; seems not to be the issue. Relanding.

Local tests with and without this patch on a Nexus 4 with and without my patch didn't show a perf regression (in fact, perf seemed to mysteriously improve -- I assume this is noise).

TBR=danakj@chromium.org
BUG=306926

Review URL: https://codereview.chromium.org/26135005