Prevent key drop if finger moves out of key area slightly before release
When typing quickly, it is quite possible that users move their fingers
slightly out of the key area before release the key. This is likely an
unintended move and the key shouldn't be dropped.

We also increased the triggering condition of swipe gesture to reduce
the accidental cursor movement.

BUG=301631

Review URL: https://codereview.chromium.org/26233005