Precache component
Layered component that precaches resources that we predict the user will want to fetch in the future, based on a list of page URLs that the user commonly visits. It interacts with a server component to determine the list of resources to precache.

This issue is a continuation of https://codereview.chromium.org/25481003, new issue created so that CL creator and CL author match.

BUG=306185

Review URL: https://codereview.chromium.org/27216002