Call WebContentsObserver::RenderViewDeleted on the RVH and possibly pending RVH when WebContents is destructed.
Currently this doesn't happen because these RVHs get destructed in RenderViewHostManager's destruction, which happens after WCO::WebContentsImplDestroyed is called. That call removes the observers from WC.

Destructing RVHM before the observers isn't practical because a lot of observers assumes they can call WC::GetRenderViewHost in their destructors. So instead, expliclity call out RenderViewDeleted.

This is needed because many of the use cases of RenderViewHostObserver is to watch for the destruction of a RenderViewHost, so this will make it less error-prone to switch to a WebContentsObserver.

BUG=306569
R=avi@chromium.org, benm@chromium.org

Review URL: https://codereview.chromium.org/27545010