Hide touch selection cursor handle when out of client view
Selection handles are made hidden when they are out of client view. This
was not done for cursor handle. Other than fixing this, a check is added
to move the handles only when they're going to be visible. Since hiding
a handle fades it out, it is better to have this fade out in the
previous handle location which is inside the boundaries of the client
view. With this change, handle positions cannot be used as selection end
points; so, whenever selection end points are needed, values cached in
touch selection controller are used instead of handle positions.

Also, IsEmptySelection() function in touch selection controller is not
needed anymore; so, it is removed.

Previous review URL: https://codereview.chromium.org/25609006/

BUG=289014

Review URL: https://codereview.chromium.org/27485008