Cache results in ExpandRectEquallyToAreaBoundedBy
About 40% of the time, PictureLayerTiling::Expand...() is called with the same arguments passed to the previous invocation on this tiling. Since this function is responsible for 5% of the time in a CPU profile of the compositor impl thread on Nexus7 (mostly because sqrt() and integer division are quite expensive), caching those results should shave away a nice slice of cost.

BUG=303123
R=danakj@chromium.org

Review URL: https://codereview.chromium.org/25555007