Fix Linux ASAN Tests (2) bot after r229536.
Direct leak of 24 byte(s) in 1 object(s) allocated from:
    #0 0x4b4c49 in operator new(unsigned long) /usr/local/google/work/chromium/src/third_party/llvm/projects/compiler-rt/lib/asan/asan_new_delete.cc:52
    #1 0x8a286e3 in base::Value::CreateStringValue(std::string const&) base/values.cc:109
    #2 0x3eba5d2 in ProfileSyncServicePreferenceTest_ModelAssociationCloudAddsOldAndNewMigratedData_Test::TestBody() chrome/browser/sync/profile_sync_service_preference_unittest.cc:531
    #3 0x7f130f7 in testing::Test::Run() testing/gtest/src/gtest.cc:2067

and so on. These showed up on the try runs of your change too, please pay
attention to this. Also, consider rewriting this code to use a RAII pattern,
then it's impossible to forget what this change adds, and the tests won't leak
when they fail.

BUG=299290
TBR=mad@chromium.org

Review URL: https://codereview.chromium.org/30633002