MetricsService: Send a hash of the UMA log in a header.
Add a new X-Chrome-UMA-Log-SHA1 header that is transmitted 
along with each UMA log upload. The header contains the SHA1 
(as hex) of the payload prior to compression. 

The hash is computed at the time that the log is closed, rather than 
just before being transmitted - to minimize the chance of hashing 
an already corrupted log. 

Does not store the hash with logs when they are persisted temporarily 
to local state for later upload, since log are already hashed when stored 
(and discarded on load if the hash does not match). The SHA1 hash is 
re-computed when logs are loaded from local state. 

BUG=308066

Review URL: https://codereview.chromium.org/26646003