Rename BeginFrame to BeginMainFrame and BeginImplFrame in cc::SchedulerStateMachine.
First patch in a series to rename BeginFrame. This patch focuses on
cc::SchedulerStateMachine.

In the scheduler, the term 'BeginFrame' is used for both the main thread and the
impl thread. This can cause confusion if it's not obvious which one is being
refered to. To clarify this, we rename BeginFrame to BeginMainFrame and
BeginImplFrame, respectively.

BUG=261765

Review URL: https://codereview.chromium.org/26969004