Remove message loop parameter from RenderThreadImpl::GetGpuFactories()
RenderThreadImpl::GetGpuFactories() makes the assumption that all the
GpuVideoAcceleratorFactories retrieved from it can use the same 3D context.
This is only true if all the factories use the same message loop for the
3D context.  Ensure this by unconditionally using
RenderThreadImpl::GetMediaThreadMessageLoopProxy() for that message loop.

This change should be a functional no-op according to the current
architecture.

BUG=306333
TEST=local build, run on CrOS snow, build on desktop Linux

Review URL: https://codereview.chromium.org/27019011