IndexedDB: Ensure that closed connections are not used
Once a database connection object is closed it doesn't
need to hold a database reference. This can be used to
detect the case where a request arrives from a renderer
after a connection is forcibly closed by the back end,
and drop the request.

In most cases the request would be associated with a
transaction that would have been aborted/removed from
the database's list and hence the request would just
be dropped. The CreateTransaction() and Close() requests
aren't, so they need this check. In the other cases 
this will just exit earlier and let the database object
be released.

BUG=308699

Review URL: https://codereview.chromium.org/26454005