[Files.app] Put a temporary action item to the file contextmenu until task is loaded
Tasks are loaded asynchronously when the context menu is requested to show. The size of menu might be changed after the menu calculates the geometry and the position so the bottom part of the menu might be hidden because the menu expanded.

This patch adds a temporary action item to the file context menu so the menu doesn't expand and a part of it is not hidden.

BUG=290390
TEST=manually tested

Review URL: https://codereview.chromium.org/27713002