Removing the avatar menu for the new multi profile style and show the avatar icons only when the window is "visiting"
With the new "multi profile" mode on ChromeOS we will have two modes:
1. Each logged in user has a separate desktop and the physical user can switch between these easily. Furthermore he can make windows "visit" the other desktop temporarily.
2. Each logged in user has his own desktop, but he will be able to see all windows from browser / applications.

In case 1 we have to show the user's icon if a window is "visiting" another desktop - but there is no avatar menu to be shown.

In case 2 we have to always show all user icons and always to show the avatar menu upon request.

This CL is addressing these two cases.

Note that the two cases will be developed in parallel to allow our users to decide what they want.

BUG=305069
TEST=visual

Review URL: https://codereview.chromium.org/26885009