[Android] MediaPlayerBridge should pass Bridge to getAllowedOperations
Right now the Java side is expecting a MediaPlayer object to be passed
to the MediaPlayerBridge.getAllowedOperations function, but native
is actually passing it a MediaPlayerBridge. Update the code to be
consistent - Java now expects the MediaPlayerBrdige and uses it to
extract a MediaPlayer.

Review URL: https://codereview.chromium.org/32213003