Allow EME base64 encoded keys to not be padded
Change the processing of EME JWK keys to add the base64 padding if it is not
present. The EME spec states that padding is not allowed, but there are some
layout tests that include padding, so this can't be enforced until the tests
are fixed.

BUG=303381

Review URL: https://codereview.chromium.org/27433002