Use adapter type rather than adapter name in chrome://diagnostics.
Before, chrome://diagnostics assumes that an interface like eth0 will be
an specific type (ie. ethernet). This may not be the case, so we explicitly
check the type of the interface. We use DebugDaemonClient::GetNetworkStatus
instead of DaebugDaemonClient::GetNetworkInterfaces so we can read the type.

BUG=239364
TEST=manual

Review URL: https://codereview.chromium.org/27590003