The inactive rendering flag on Windows AURA should be turned off in the WM_NCACTIVATE handler when the window is activated.
Currently this is turned off when the aura Window is deactivated, which occurs when a top level window like a bubble is
displayed. This causes the browser frame's caption to be displayed as inactive when a bubble is displayed which is not
correct.

Removed the NativeWidgetAuraWindowObserver class as it does not do anything useful with the inactive rendering flag setting
code taken out. I also removed the SetInactiveRenderingDisabled method from the NativeWidgetPrivate interface as it does
not do anything useful.

The inactive rendering flag gets reset correctly in the WM_ACTIVATEAPP and WM_NCACTIVATE messages on Windows.

BUG=307343
R=sky@chromium.org, sky

Review URL: https://codereview.chromium.org/26800010