Move evdev events support into EventFactoryDelegateEvdev
And introduce a gyp variable use_ozone_evdev to allow this feature to be
removed from the build. Additionally, the --ozone-events flag is introduced
to select an EventFactoryDelegateOzone implementation to use.

If an embedder or out-of-tree EventFactoryDelegateOzone implementation
calls SetEventFactoryDelegateOzone during initialization, the flag has no
effect.

BUG=none
TEST=Build & run content_shell, ui_unittests with use_ozone=1

Review URL: https://codereview.chromium.org/28513004