The current code in cast_socket_api would possibly trigger deletion of the CastSocket during a callback invoked by the CastSocket, by removing it from the APIResourceManager.
This patch:

(1) Always adds the socket to the resource manager on construction (which takes ownership)

(2) Eliminates the step to remove the socket from the resource manager on error

Thus making it incumbent on the client to call Close() to dispose of the socket.

TODO: Still need to understand when it is safe to remove something from the resource manager.

BUG=295843

Review URL: https://codereview.chromium.org/26972006