[chromium_android] Use zip to archive build products instead of tar.
This CL also removes the 'archive' recipes module which would end up
without coverage otherwise. The 'zip' and 'unzip' functionalities
are got via the shell commands 'zip' and 'unzip' respectively.

The reason this CL is required is because the existing 'tar'
functionality comes in via the Python module 'tarfile' which
is extremely slow and is timing out of few bots.

BUG=286509

Review URL: https://codereview.chromium.org/32323010