Add helper method to delete and remove SourceBufferRange objects.
Consolidates logic for deleting a SourceBufferRange and removing
it from the ranges_ list. This change will make it easier to ensure
that cached references to ranges, like selected_range_, always
get cleared when a range is removed. No changes in functionality
are included in this patch.

BUG=303413
TEST=All existing unit tests still pass.
R=scherkus@chromium.org

Review URL: https://codereview.chromium.org/29453002