Add presubmit check to catch use of PRODUCT_NAME in resources.
We should not be adding any new strings resources that use a PRODUCT_NAME
placeholder, because it is difficult to localize in some languages. Add a
presubmit check which shows a warning whenever a line is that contains
PRODUCT_NAME is changed in a .grd file.

BUG=None

Review URL: https://codereview.chromium.org/32213002