Disable linux_swarm_triggered.
It was enabled in r230165 (a few moments ago, really) but we want to do
infrastructure work on the swarming master before increasing usage. This
CL will be reverted soon, we simply do not want the CQ to be restarted
in the meantime and having it start using linux_swarm_triggered while
maintenance is going on.

TBR=stip@chromium.org
BUG=

Review URL: https://codereview.chromium.org/35813002