android: Use new proxy from PROXY_CHANGE intent
When we get a PROXY_CHANGE intent, the intent comes with the new proxy value;
use this value instead of ignoring it and asking the system property store for
the new proxy. This fixes a race condition where the system property store may
not have the new proxy by the time we receive the intent.

BUG=none yet
TEST=trybot

Review URL: https://codereview.chromium.org/26763005