Track Welcome Notification Pop Up State
To prevent continuously popping up the welcome notification on each
Chrome startup, this delta adds a local preference to track if
we popped up the notification. If we have to show the notification and
we already popped the notification, show it in a non-popped state.

BUG=305449

Review URL: https://codereview.chromium.org/33023003