Improve Windows GN build setup.
Properly setup a new output build directory on Windows (previously you had to manually copy some files into it). It now creates the environment block and the gyp-win-tool.

The include and library paths are moved to the command line and removed from the environment which is cleaner and more explicit.

BUG=
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/34603009