linux_aura: Enable interactive_ui_tests.
Previously, the chrome test runner had a check compiled in only in linux_aura where it detected if the name of the binary was "interactive_ui_tests" and quit immediately.

This removes that check, provides a minimal implementation of ui controls for desktop aura, and enables 56 tests.

BUG=163931

Review URL: https://codereview.chromium.org/26315005