Allow asynchronous GPU context creation for browser clients
This moves the call to establish a GPU channel out of WGC3D.
It adds an asynchronous method to BrowserGpuChannelHostFactory
that registers callbacks that are run on the main thread once
the channel is established.

This is prerequisite work for asynchronous OutputSurface creation.

BUG=270179,125248,235300
R=piman@chromium.org

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=228777

Review URL: https://codereview.chromium.org/26959002