sync: Refactor update verification and processing
Define a new function for partitioning an update response into per-type
chunks and use it to refactor ProcessUpdateCommand's functions.

This is part of the preparation to move towards maintaining a set of
per-type objects that we can use to process updates in different ways
for different types.

This commit also updates the GetModelTypeFromSpecifics() so it no longer
reports the type of deleted items as "UNSPECIFIED".  The old behavior
was implemented in a time when the server would not set the type of
tombstone items.  The current server will always populate that
information.

BUG=278484

Review URL: https://codereview.chromium.org/27765002